+++ composeApp/src/commonMain/kotlin/com/mouse/mouse/core/domain/model/MorseMode.kt
@@ -1,6 +1,0 @@
-package com.mouse.mouse.core.domain.model
-
-enum class MorseMode {
-    TEXT_TO_MORSE,
-    MORSE_TO_TEXT
-}