+++ composeApp/src/commonMain/kotlin/com/mouse/mouse/core/domain/model/PlaybackState.kt
@@ -1,8 +1,0 @@
-package com.mouse.mouse.core.domain.model
-
-enum class PlaybackState {
-    IDLE,
-    PREPARING,
-    PLAYING,
-    ERROR
-}